backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
motor>=3.6.0
pydantic>=2.10.0
pydantic-settings>=2.6.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
slowapi>=0.1.9
httpx>=0.28.0
google-generativeai>=0.8.3
python-multipart>=0.0.17
email-validator>=2.2.0
magic-admin>=2.2.0
prometheus-client>=0.21.0
pytest>=8.3.4
pytest-asyncio>=0.24.0
